refactor(PI-8): Refactorización de procesadores HTML y Markdown
Se ha eliminado la clase `HtmlDocumentProcessor` de `HtmlDocumentProcessor.cs`.
Se ha añadido una nueva clase `HtmlProcessor` en `HtmlProcessor.cs` para manejar el procesamiento de documentos HTML. Esta clase utiliza `MarkdownProcessor` para procesar el contenido convertido a markdown.
La clase `HtmlProcessor` convierte el contenido HTML a markdown usando la biblioteca `ReverseMarkdown` y luego limpia el contenido markdown antes de pasarlo a `MarkdownProcessor`.
Se ha añadido una nueva clase `MarkdownProcessor` en `MarkdownProcessor.cs` para manejar el procesamiento de documentos markdown. Esta clase divide el contenido markdown y crea objetos `Document` para cada fragmento.
Se ha actualizado el archivo `ServicesRegistration.cs` para registrar las nuevas clases `HtmlProcessor` y `MarkdownProcessor` en el contenedor de inyección de dependencias. `HtmlProcessor` se registra para las extensiones de archivo `.htm` y `.html`, y `MarkdownProcessor` se registra para las extensiones de archivo `.md`.
La clase `HtmlProcessor` incluye un método `CleanMarkdown` para limpiar el contenido markdown normalizando listas, eliminando citas en bloque, transformando títulos de tablas en encabezados y eliminando contenido de pie de página.
La clase `MarkdownProcessor` incluye métodos para dividir el contenido markdown, extraer texto de nodos y encabezados, y construir fragmentos a partir de encabezados y texto.
